Portability fixes found by turning on compiler warnings.
These fall into the following major groups:

 - Uninitialized variables. Several could definitely be left uninitialized
   in normal use, and the code was assuming they would be zero-filled.
 - Missing #includes. Minor, but found bugs.
 - Nonportable functions, notably sysinfo(), which was being called in
   several places even though it is completely different on Solaris and
   Linux. On Linux, uname() does the job.
 - Casts of types whose size can vary, mostly in printf() and sprintf()
   calls.
 - other (aliasing violations, compilation-breaking typos, broken error
   checks of e.g. fopen(), uninitialized accesses due to #if 0 in the
   wrong place).

Also remove some unused variable declarations and dead portability
functions (notably fork1()), and decorate functions that can plainly
not return with a _dt_noreturn_, newly defined in <compiler.h> to
equal __attribute__((__noreturn__)).

Finally, this actually turns on compiler warnings, and moves to
LARGEFILE64 mode (which the code is plainly meant to run in, as
it calls open64()).

A few possible bugs are just FIXMEd, because I am not yet sure if the
problematic situation can ever occur.

+++ libdtrace/dt_consume.c
@@ -34,6 +34,7 @@
 #include <alloca.h>
 #include <dt_impl.h>
 #include <libproc.h>
+#include <port.h>
 
 #define	DT_MASK_LO 0x00000000FFFFFFFFULL
 
@@ -870,7 +871,7 @@ dt_print_stack(dtrace_hdl_t *dtp, FILE *fp, const char *format,
 			if (pc > sym.st_value) {
 				(void) snprintf(c, sizeof (c), "%s`%s+0x%llx",
 				    dts.dts_object, dts.dts_name,
-				    pc - sym.st_value);
+				    (long long unsigned) pc - sym.st_value);
 			} else {
 				(void) snprintf(c, sizeof (c), "%s`%s",
 				    dts.dts_object, dts.dts_name);
@@ -883,9 +884,10 @@ dt_print_stack(dtrace_hdl_t *dtp, FILE *fp, const char *format,
 			 */
 			if (dtrace_lookup_by_addr(dtp, pc, NULL, &dts) == 0) {
 				(void) snprintf(c, sizeof (c), "%s`0x%llx",
-				    dts.dts_object, pc);
+				    dts.dts_object, (long long unsigned) pc);
 			} else {
-				(void) snprintf(c, sizeof (c), "0x%llx", pc);
+				(void) snprintf(c, sizeof (c), "0x%llx",
+				    (long long unsigned) pc);
 			}
 		}
